@sadikovi 2023-08-29
I could only confirm this regression and the fix in the internal benchmarks, it could take quite some time to port the benchmark + data generation to open source. Is it possible to add it as a follow-up?