@sadikovi 2023-08-31
@MaxGekk @cloud-fan Can you review again? 
I have added the benchmark that shows performance almost 3x performance improvement.