@sadikovi 2023-09-04
Shall we merge this or do you have any concerns or questions? I will be more than happy to answer them or follow up on the suggestions.

We may also need to backport to Spark 3.5/3.4 where this change was originally introduced, I can open a separate PR for the backport.