@sadikovi 2023-09-04
Great, thanks. I will follow up on your comment regarding the possible regression in the parsing benchmark.